inserita sezione trattamento e cambio stato appuntamento automatico da finire trattamento

+++ frontend/src/app/graphql/operations/availability-appointment.queries.ts
@@ -117,8 +117,8 @@ export const IS_INSTRUMENT_AVAILABLE = gql`
     $instrumentId: ID!
     $appointmentDate: String!
     $startTime: String!
-    $startOffsetMinutes: Int!
-    $endOffsetMinutes: Int!
+    $startOffsetMinutes: Float!
+    $endOffsetMinutes: Float!
     $excludeAppointmentId: ID
   ) {
     isInstrumentAvailable(
